requirements.txt
numpy
scipy
websocket-client
yt-dlp>=2024.12.23
ffmpeg-python>=0.2.0,<0.3
sounddevice<1.0
openai-whisper==20240930
faster-whisper>=1.1.0,<2.0.0
openai>=1.40,<2.0
google-generativeai>=0.7,<1.0
